started lipa mdogo mdogo and fixed category fetch and creation error

+++ src/main/java/org/example/qposbackend/DTOs/StockItemDTO.java
@@ -1,5 +1,5 @@
 package org.example.qposbackend.DTOs;
 
 public record StockItemDTO(Long item, Double buyingPrice,  Integer packaging,
-                           Integer quantity) {
+                           Double quantity) {
 }
