+++ src/main/java/org/example/qposbackend/InventoryItem/quantityAdjustment/dto/QuantityAdjustmentDto.java
@@ -1,3 +1,3 @@
 package org.example.qposbackend.InventoryItem.quantityAdjustment.dto;
 
-public record QuantityAdjustmentDto(Integer quantity, String reason){}
+public record QuantityAdjustmentDto(Double quantity, String reason){}
